Update
Increased radius coverage to 3 miles from 1.  As well as adding a
checker to the weighted average formula to ensure that, if for example
we have a result set of size 1 or multiple entries that have the same
house price we would be dividing by 0. So we would get an error.